requirements.txt
django
django-bootstrap3
requests
django-multiselectfield
bson
django-allauth